+++ src/main/java/org/spongepowered/common/mixin/core/block/MixinBlockCrops.java
@@ -77,7 +77,7 @@ public abstract class MixinBlockCrops extends MixinBlock {
     }
 
     private ImmutableGrowthData getGrowthData(IBlockState blockState) {
-        return ImmutableDataCachingUtil.getManipulator(ImmutableSpongeGrowthData.class, (Integer) blockState.getValue(BlockCrops.AGE), 0, 7);
+        return ImmutableDataCachingUtil.getManipulator(ImmutableSpongeGrowthData.class, blockState.getValue(BlockCrops.AGE), 0, 7);
     }
 
 }
